ff-29: clean up controller methods into services
