+++ src/main/java/com/flagfinder/service/GameNameService.java
@@ -1,0 +1,11 @@
+package com.flagfinder.service;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+
+import java.util.Map;
+
+public interface GameNameService {
+    ResponseEntity<?> setGameName(Map<String, String> request, Authentication authentication);
+    ResponseEntity<?> checkGameNameAvailability(String gameName);
+}
